requirements.txt
asgiref==3.8.1
azure-core==1.34.0
azure-storage-blob==12.25.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.0.14
django-cors-headers==4.7.0
django-mssql-backend==2.8.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
mssql-django==1.5
mysqlclient==2.2.7
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyodbc==5.2.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
whitenoise==6.9.0
